+++ Questionable/Controller/Steps/Shared/AbandonQuest.cs
@@ -1,14 +1,15 @@
 ﻿using System.Globalization;
 using Dalamud.Game.ClientState.Conditions;
 using Dalamud.Plugin.Services;
+using ECommons.Throttlers;
 using FFXIVClientStructs.FFXIV.Client.Game;
 using Microsoft.Extensions.Logging;
-using ECommons.Throttlers;
+using Questionable.Data;
 using Questionable.Functions;
 using Questionable.Model;
-using Questionable.Data;
 
 namespace Questionable.Controller.Steps.Shared;
+
 internal static class AbandonQuest
 {
 
@@ -55,7 +56,7 @@ internal static class AbandonQuest
             {
                 throw new TaskException("Quest cannot be cancelled");
             }
-            
+
             AbandonQuestAction();
             return true;
         }